@@ -15,17 +15,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.1",
-    "@angular/cdk": "~13.0.1",
-    "@angular/common": "~13.0.1",
-    "@angular/compiler": "~13.0.1",
-    "@angular/core": "~13.0.1",
-    "@angular/forms": "~13.0.1",
-    "@angular/material": "^13.0.1",
-    "@angular/platform-browser": "~13.0.1",
-    "@angular/platform-browser-dynamic": "~13.0.1",
-    "@angular/router": "~13.0.1",
-    "@angular/service-worker": "~13.0.1",
+    "@angular/animations": "^16.1.3",
+    "@angular/cdk": "^16.1.3",
+    "@angular/common": "^16.1.3",
+    "@angular/compiler": "^16.1.3",
+    "@angular/core": "^16.1.3",
+    "@angular/forms": "^16.1.3",
+    "@angular/material": "^16.1.3",
+    "@angular/platform-browser": "^16.1.3",
+    "@angular/platform-browser-dynamic": "^16.1.3",
+    "@angular/router": "^16.1.3",
+    "@angular/service-worker": "^16.1.3",
     "@apidevtools/json-schema-ref-parser": "^9.0.7",
     "@apollo/client": "^3.0.0",
     "@auth0/angular-jwt": "^5.0.2",
@@ -33,17 +33,17 @@
     "@fullcalendar/daygrid": "4.1.0",
     "@fullcalendar/interaction": "4.1.0",
     "@fullcalendar/timegrid": "4.1.0",
-    "@ngrx/store": "^12.5.1",
-    "@ngrx/store-devtools": "^12.5.1",
-    "@ngx-translate/core": "^13.0.0",
+    "@ngrx/store": "^16.0.1",
+    "@ngrx/store-devtools": "^16.0.1",
+    "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "@stoplight/json-ref-resolver": "^3.1.2",
-    "apollo-angular": "^2.6.0",
+    "apollo-angular": "^5.0.0",
     "chart.js": "^3.9.1",
-    "flag-icon-css": "^3.5.0",
-    "graphql": "^15.0.0",
+    "flag-icons": "^6.7.0",
+    "graphql": "^16.7.1",
     "lodash-es": "^4.17.21",
-    "ng-recaptcha": "^8.0.1",
+    "ng-recaptcha": "^12.0.1",
     "ngx-take-until-destroy": "^5.4.0",
     "primeflex": "^2.0.0",
     "primeicons": "^4.1.0",
@@ -51,27 +51,27 @@
     "pwa": "^1.9.7",
     "rxjs": "~6.6.0",
     "tslib": "^2.1.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^13.0.0",
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-eslint/builder": "12.6.1",
-    "@angular-eslint/eslint-plugin": "12.6.1",
-    "@angular-eslint/eslint-plugin-template": "12.6.1",
-    "@angular-eslint/schematics": "12.6.1",
-    "@angular-eslint/template-parser": "12.6.1",
-    "@angular/cli": "~13.0.2",
-    "@angular/compiler-cli": "~13.0.1",
-    "@ngrx/schematics": "^12.5.1",
+    "@angular-devkit/build-angular": "^16.1.3",
+    "@angular-eslint/builder": "16.0.3",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/schematics": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
+    "@angular/cli": "^16.1.3",
+    "@angular/compiler-cli": "^16.1.3",
+    "@ngrx/schematics": "^16.0.1",
     "@types/jasmine": "~3.6.0",
     "@types/lodash-es": "^4.17.4",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.32.0",
-    "@typescript-eslint/parser": "4.32.0",
-    "eslint": "^7.32.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
+    "eslint": "^8.39.0",
     "eslint-plugin-graphql": "^4.0.0",
-    "eslint-plugin-ngrx": "^1.46.3",
+    "eslint-plugin-ngrx": "^2.1.4",
     "eslint-plugin-rxjs": "^3.3.7",
     "husky": "^8.0.0",
     "jasmine-core": "~3.7.0",
@@ -83,7 +83,7 @@
     "ng2-charts": "^3.1.2",
     "prettier": "^2.8.7",
     "pretty-quick": "^3.1.3",
-    "typescript": "~4.4.4",
+    "typescript": "~5.1.6",
     "webpack-bundle-analyzer": "^4.4.1",
     "webpack-sources": "^3.2.1"
   }
