@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { MusicianProfileDto } from '@arpa/models';
 import { first } from 'rxjs/operators';
 import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
@@ -12,18 +12,17 @@ import { NotificationsService } from '@arpa/services';
   styleUrls: ['./musician-deactivation.component.scss'],
 })
 export class MusicianDeactivationComponent implements OnInit {
-
-  public form: FormGroup;
+  public form: UntypedFormGroup;
 
   public profile: MusicianProfileDto;
   public minDate: Date;
 
   constructor(
     public config: DynamicDialogConfig,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private musicianService: MusicianService,
     public ref: DynamicDialogRef,
-    private notificationsService: NotificationsService,
+    private notificationsService: NotificationsService
   ) {
     this.config.data.profile.pipe(first()).subscribe((profile: MusicianProfileDto) => {
       this.profile = profile;
@@ -48,7 +47,8 @@ export class MusicianDeactivationComponent implements OnInit {
   }
 
   activate() {
-    this.musicianService.activateProfile(this.profile.id)
+    this.musicianService
+      .activateProfile(this.profile.id)
       .pipe(first())
       .subscribe(() => {
         this.notificationsService.success('ACTIVATED', 'musician-profile-dialog');
@@ -60,12 +60,12 @@ export class MusicianDeactivationComponent implements OnInit {
     if (this.form.invalid || this.form.pristine) {
       return;
     }
-    this.musicianService.deactivateProfile(this.profile.id, { ...this.form.value })
+    this.musicianService
+      .deactivateProfile(this.profile.id, { ...this.form.value })
       .pipe(first())
       .subscribe((result) => {
         this.notificationsService.success('DEACTIVATED', 'musician-profile-dialog');
         this.ref.close();
       });
   }
-
 }
