@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { first, map } from 'rxjs/operators';
 import { SelectValueService, NotificationsService } from '@arpa/services';
 import { MusicianProfileDto, EducationDto } from '@arpa/models';
@@ -15,7 +15,7 @@ import { ColumnDefinition } from '../../../../@arpa/components/table/table.compo
   styleUrls: ['./musician-education.component.scss'],
 })
 export class MusicianEducationComponent implements OnInit {
-  public form: FormGroup;
+  public form: UntypedFormGroup;
 
   public profile: MusicianProfileDto;
 
@@ -29,7 +29,7 @@ export class MusicianEducationComponent implements OnInit {
 
   constructor(
     public config: DynamicDialogConfig,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     public ref: DynamicDialogRef,
     private selectValueService: SelectValueService,
     private musicianService: MusicianService,
