@@ -1,6 +1,6 @@
 import { EnumService } from './../../../../@arpa/services/enum.service';
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { SetMyAppointmentParticipationPredictionDto } from '@arpa/models';
 import { Observable } from 'rxjs';
 import { SelectItem } from 'primeng/api';
@@ -9,19 +9,14 @@ import { DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
 @Component({
   selector: 'arpa-my-appointment-participation-dialog',
   templateUrl: './my-appointment-participation-dialog.component.html',
-  styleUrls: ['./my-appointment-participation-dialog.component.scss']
+  styleUrls: ['./my-appointment-participation-dialog.component.scss'],
 })
 export class MyAppointmentParticipationDialogComponent implements OnInit {
-
-  form: FormGroup;
+  form: UntypedFormGroup;
   participation: SetMyAppointmentParticipationPredictionDto = this.config.data.participation;
   statusOptions$: Observable<SelectItem[]> = this.config.data.statusOptions$;
 
-  constructor(
-    private formBuilder: FormBuilder,
-    public config: DynamicDialogConfig,
-    public ref: DynamicDialogRef
-  ) {}
+  constructor(private formBuilder: UntypedFormBuilder, public config: DynamicDialogConfig, public ref: DynamicDialogRef) {}
 
   ngOnInit() {
     this.form = this.formBuilder.group({
@@ -42,5 +37,4 @@ export class MyAppointmentParticipationDialogComponent implements OnInit {
   cancel() {
     this.ref.close(null);
   }
-
 }
