@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { BehaviorSubject } from 'rxjs';
 import { ColumnDefinition } from '../../../../@arpa/components/table/table.component';
 import { BankAccountDto, PersonDto } from '@arpa/models';
@@ -10,33 +10,24 @@ import { cloneDeep } from 'lodash-es';
 @Component({
   selector: 'arpa-user-bankdata',
   templateUrl: './user-bankdata.component.html',
-  styleUrls: ['./user-bankdata.component.scss']
+  styleUrls: ['./user-bankdata.component.scss'],
 })
 export class UserBankdataComponent implements OnInit {
-
-  public form: FormGroup;
+  public form: UntypedFormGroup;
   tableData: BehaviorSubject<any> = new BehaviorSubject([]);
   private _tableData: Array<any>;
   @Input() bankAccounts: BankAccountDto[] = [];
   @Input() person: PersonDto | null;
-
-
 
   columns: ColumnDefinition<BankAccountDto>[] = [
     { label: 'profile.bank.IBAN', property: 'iban', type: 'text' },
     { label: 'profile.bank.BIC', property: 'bic', type: 'text' },
     { label: 'profile.bank.ACCOUNT_OWNER', property: 'accountOwner', type: 'text' },
     { label: 'profile.bank.COMMENT_INNER', property: 'commentInner', type: 'text' },
-    { label: 'profile.bank.STATUS', property: 'statusId', type: 'state', stateTable: 'BankAccount', stateProperty: 'Status'},
-
-
+    { label: 'profile.bank.STATUS', property: 'statusId', type: 'state', stateTable: 'BankAccount', stateProperty: 'Status' },
   ];
 
-  constructor(
-    private formBuilder: FormBuilder,
-    private meService: MeService,
-    private notificationsService: NotificationsService,
-  ) {
+  constructor(private formBuilder: UntypedFormBuilder, private meService: MeService, private notificationsService: NotificationsService) {
     this.form = this.formBuilder.group({
       commentInner: [null, [Validators.maxLength(500)]],
       id: [null],
@@ -51,7 +42,6 @@ export class UserBankdataComponent implements OnInit {
     this._tableData = this.bankAccounts && this.bankAccounts.length ? cloneDeep(this.bankAccounts) : [];
     this.tableData.next(this._tableData);
   }
-
 
   onSubmit() {
     const { id, commentInner, iban, bic, accountOwner } = this.form.getRawValue();
@@ -79,7 +69,6 @@ export class UserBankdataComponent implements OnInit {
     }
   }
 
-
   remove(bankAccount: BankAccountDto): void {
     this.meService
       .deleteBankAccount(bankAccount.id, this.person?.id)
@@ -89,7 +78,6 @@ export class UserBankdataComponent implements OnInit {
         this.notificationsService.success('BANK_ACCOUNT_REMOVED', 'bank');
       });
   }
-
 
   update(bankAccounts: BankAccountDto) {
     this.form.reset({
