@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { ActivatedRouteSnapshot, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
 import { NotificationsService } from '../services/notifications.service';
 import { map } from 'rxjs/operators';
@@ -8,16 +8,15 @@ import { AuthService } from '../services/auth.service';
 @Injectable({
   providedIn: 'root',
 })
-export class AuthGuard implements CanActivate {
-  constructor(private authService: AuthService, private router: Router, private notificationsService: NotificationsService) {
-  }
+export class AuthGuard {
+  constructor(private authService: AuthService, private router: Router, private notificationsService: NotificationsService) {}
 
   canActivate(
     route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-
+    state: RouterStateSnapshot
+  ): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
     return this.authService.isAuthenticated.pipe(
-      map(loggedIn => {
+      map((loggedIn) => {
         if (!loggedIn) {
           this.notificationsService.info('error.NOT_AUTHENTICATED');
           this.router.navigate(['/login']);
@@ -25,8 +24,7 @@ export class AuthGuard implements CanActivate {
         }
 
         return true;
-      }),
+      })
     );
-
   }
 }
