@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivateChild, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { ActivatedRouteSnapshot, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
 import { AuthService } from '../services/auth.service';
 import { map } from 'rxjs/operators';
@@ -7,15 +7,13 @@ import { map } from 'rxjs/operators';
 @Injectable({
   providedIn: 'root',
 })
-export class RoleGuard implements CanActivateChild {
-
-  constructor(private router: Router, private authService: AuthService) {
-  }
+export class RoleGuard {
+  constructor(private router: Router, private authService: AuthService) {}
 
   canActivateChild(
     route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-
+    state: RouterStateSnapshot
+  ): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
     const requiredRoles = route.data.roles;
 
     if (!requiredRoles) {
@@ -28,8 +26,7 @@ export class RoleGuard implements CanActivateChild {
           return true;
         }
         return this.router.parseUrl('/forbidden');
-      }),
+      })
     );
   }
-
 }
