@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivateChild, Router, UrlTree } from '@angular/router';
+import { ActivatedRouteSnapshot, Router, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
 import { AuthService } from '../services/auth.service';
 import { map } from 'rxjs/operators';
@@ -7,19 +7,17 @@ import { map } from 'rxjs/operators';
 @Injectable({
   providedIn: 'root',
 })
-export class SessionGuard implements CanActivateChild {
-  constructor(private authService: AuthService, private router: Router) {
+export class SessionGuard {
+  constructor(private authService: AuthService, private router: Router) {}
+
+  canActivateChild(childRoute: ActivatedRouteSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    return this.authService.isAuthenticated.pipe(
+      map((isAuthenticated) => {
+        if (isAuthenticated && childRoute.data.sessionPrevent) {
+          this.router.navigate(['/arpa']);
+        }
+        return true;
+      })
+    );
   }
-
-  canActivateChild(childRoute: ActivatedRouteSnapshot):
-    Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-
-    return this.authService.isAuthenticated.pipe(map((isAuthenticated) => {
-      if (isAuthenticated && childRoute.data.sessionPrevent) {
-        this.router.navigate(['/arpa']);
-      }
-      return true;
-    }));
-  }
-
 }
