@@ -1,13 +1,12 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve } from '@angular/router';
+import { ActivatedRouteSnapshot } from '@angular/router';
 import { Observable, of } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { SectionService } from '@arpa/services';
 
 @Injectable()
-export class SectionListResolver implements Resolve<boolean> {
-  constructor(private sectionService: SectionService) {
-  }
+export class SectionListResolver {
+  constructor(private sectionService: SectionService) {}
 
   resolve(route: ActivatedRouteSnapshot): Observable<boolean> {
     if (this.sectionService.sectionsLoaded) {
