@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { ActivatedRouteSnapshot, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { AuthService } from '../../../@arpa/services/auth.service';
@@ -7,15 +7,13 @@ import { AuthService } from '../../../@arpa/services/auth.service';
 @Injectable({
   providedIn: 'root',
 })
-export class DashboardRoleGuard implements CanActivate {
-
-  constructor(private router: Router, private authService: AuthService) {
-  }
+export class DashboardRoleGuard {
+  constructor(private router: Router, private authService: AuthService) {}
 
   canActivate(
     route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-
+    state: RouterStateSnapshot
+  ): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
     let { role: roleName } = route.params;
 
     return this.authService.isUserInAtLeastOnRole([roleName]).pipe(
@@ -24,7 +22,7 @@ export class DashboardRoleGuard implements CanActivate {
           return true;
         }
         return this.router.parseUrl('/forbidden');
-      }),
+      })
     );
   }
 }
