@@ -21,11 +21,11 @@ import { InputTextModule } from 'primeng/inputtext';
 import { LanguageService } from '@arpa/services';
 import { LayoutModule } from '@angular/cdk/layout';
 import { LoadingModule } from '../../../@arpa/components/loading/loading.module';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
+import { MatLegacyButtonModule as MatButtonModule } from '@angular/material/legacy-button';
+import { MatLegacyCardModule as MatCardModule } from '@angular/material/legacy-card';
 import { MatGridListModule } from '@angular/material/grid-list';
 import { MatIconModule } from '@angular/material/icon';
-import { MatMenuModule } from '@angular/material/menu';
+import { MatLegacyMenuModule as MatMenuModule } from '@angular/material/legacy-menu';
 import { MessagesWidgetComponent } from './messages-widget/messages-widget.component';
 import { NgModule } from '@angular/core';
 import { OverlayPanelModule } from 'primeng/overlaypanel';
