@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { MusicianProfileDto } from '@arpa/models';
 import { MuproService } from '../services/mupro.service';
@@ -7,7 +7,7 @@ import { MuproService } from '../services/mupro.service';
 @Injectable({
   providedIn: 'root',
 })
-export class MusicianProfileResolver implements Resolve<MusicianProfileDto> {
+export class MusicianProfileResolver {
   constructor(private muproService: MuproService) {}
 
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<MusicianProfileDto> {
