@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { MeService } from '@arpa/services';
 import { MusicianProfileDto, MyMusicianProfileDto } from '@arpa/models';
@@ -8,9 +8,7 @@ import { MusicianService } from '../services/musician.service';
 @Injectable({
   providedIn: 'root',
 })
-export class ProfilesResolver
-  implements Resolve<MyMusicianProfileDto | MyMusicianProfileDto[] | MusicianProfileDto | MusicianProfileDto[]>
-{
+export class ProfilesResolver {
   constructor(private meService: MeService, private musicianService: MusicianService) {}
 
   resolve(
