@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { MeService } from '@arpa/services';
 import { MyMusicianProfileDto } from '@arpa/models';
@@ -7,7 +7,7 @@ import { MyMusicianProfileDto } from '@arpa/models';
 @Injectable({
   providedIn: 'root',
 })
-export class ProfileMusicianResolver implements Resolve<MyMusicianProfileDto | MyMusicianProfileDto[]> {
+export class ProfileMusicianResolver {
   constructor(private meService: MeService) {}
 
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<MyMusicianProfileDto | MyMusicianProfileDto[]> {
