@@ -1,13 +1,12 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { MeService } from '@arpa/services';
 import { MyUserProfileDto } from '@arpa/models';
 
 @Injectable()
-export class ProfileResolver implements Resolve<MyUserProfileDto> {
-  constructor(private meService: MeService) {
-  }
+export class ProfileResolver {
+  constructor(private meService: MeService) {}
 
   public resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<MyUserProfileDto> {
     return this.meService.getMyProfile();
