@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { SectionService } from '@arpa/services';
 import { SectionDto } from '@arpa/models';
@@ -7,9 +7,8 @@ import { SectionDto } from '@arpa/models';
 @Injectable({
   providedIn: 'root',
 })
-export class SectionsResolver implements Resolve<SectionDto[]> {
-  constructor(private sectionService: SectionService) {
-  }
+export class SectionsResolver {
+  constructor(private sectionService: SectionService) {}
 
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<SectionDto[]> {
     return this.sectionService.load();
