@@ -20,8 +20,8 @@ import { LanguageService } from '@arpa/services';
 import { GraphQlFeedModule } from '../../../@arpa/components/graph-ql-feed/graph-ql-feed.module';
 import { TableModule } from '../../../@arpa/components/table/table.module';
 import { FormFieldModule } from '../../../@arpa/components/form-field/form-field.module';
-import { MatMenuModule } from '@angular/material/menu';
-import { MatButtonModule } from '@angular/material/button';
+import { MatLegacyMenuModule as MatMenuModule } from '@angular/material/legacy-menu';
+import { MatLegacyButtonModule as MatButtonModule } from '@angular/material/legacy-button';
 import { MatIconModule } from '@angular/material/icon';
 import { ProjectLayoutComponent } from './project-layout/project-layout.component';
 import { TabViewModule } from 'primeng/tabview';
